test(test-optimization): run Vitest context tests in CI
