+++ packages/datadog-instrumentations/test/vitest-main.spec.js
@@ -5,6 +5,37 @@ const assert = require('node:assert/strict')
 const proxyquire = require('proxyquire').noPreserveCache()
 
 const { EMPTY_EFD_RETRY_POLICY } = require('../../dd-trace/src/ci-visibility/efd-retry-policy')
+const {
+  makeProvidedContextBrowserSafe,
+  parseProvidedContextValue,
+} = require('../src/vitest-util')
+
+describe('vitest utilities', () => {
+  describe('browser-safe provided context', () => {
+    it('leaves safe context unchanged', () => {
+      const context = { knownTests: ['safe test'] }
+
+      assert.strictEqual(makeProvidedContextBrowserSafe(context), context)
+      assert.strictEqual(parseProvidedContextValue(context), context)
+    })
+
+    it('round trips context containing a closing script tag without exposing HTML markup', () => {
+      const context = {
+        knownTests: ['test containing </ScRiPt> and <markup>'],
+      }
+
+      const safeContext = makeProvidedContextBrowserSafe(context)
+
+      assert.strictEqual(typeof safeContext, 'string')
+      assert.ok(!safeContext.includes('<'))
+      assert.deepStrictEqual(parseProvidedContextValue(safeContext), context)
+    })
+
+    it('rejects malformed serialized context', () => {
+      assert.strictEqual(parseProvidedContextValue('{'), undefined)
+    })
+  })
+})
 
 describe('vitest main instrumentation', () => {
   it('keeps no-worker capabilities active after Browser Mode setup', async () => {
